homework08/requirements/base.txt
Django==2.1.3
psycopg2-binary==2.7.6
python-decouple==3.1
pytz==2018.3
uWSGI==2.0.15
